Moved buffer transferring logic under the self renderer
