+++ Glade2d/Graphics/SelfRenderer/BufferTransferring/NoRotationBufferTransferrer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
-using Glade2d.Graphics.SelfRenderer;
 using Meadow.Foundation.Graphics.Buffers;
 
-namespace Glade2d.Graphics.BufferTransferring;
+namespace Glade2d.Graphics.SelfRenderer.BufferTransferring;
 
-internal class Rotation180BufferTransferrer : IBufferTransferrer
+internal class NoRotationBufferTransferrer : IBufferTransferrer
 {
     public void Transfer(BufferRgb565 source, BufferRgb565 target, int scale)
     {
@@ -28,14 +27,13 @@ internal class Rotation180BufferTransferrer : IBufferTransferrer
             fixed (byte* sourceBufferPtr = source.Buffer)
             fixed (byte* targetBufferPtr = target.Buffer)
             {
-                var targetRowStartIndex = (targetHeight - 1) * targetWidth * GladeSelfRenderer.BytesPerPixel;
+                var targetRowStartIndex = 0;
                 var sourceByte1 = sourceBufferPtr;
-                var targetByte1 = targetBufferPtr +
-                                  targetWidth * targetHeight * GladeSelfRenderer.BytesPerPixel -
-                                  GladeSelfRenderer.BytesPerPixel;
+                var targetByte1 = targetBufferPtr;
 
                 for (var sourceRow = 0; sourceRow < sourceHeight; sourceRow++)
                 {
+                    // First copy the source row scaled horizontally
                     for (var sourceCol = 0; sourceCol < sourceWidth; sourceCol++)
                     {
                         for (var scaleX = 0; scaleX < scale; scaleX++)
@@ -43,26 +41,27 @@ internal class Rotation180BufferTransferrer : IBufferTransferrer
                             *targetByte1 = *sourceByte1;
                             *(targetByte1 + 1) = *(sourceByte1 + 1);
 
-                            targetByte1 -= GladeSelfRenderer.BytesPerPixel;
+                            targetByte1 += GladeSelfRenderer.BytesPerPixel;
                         }
 
                         sourceByte1 += GladeSelfRenderer.BytesPerPixel;
                     }
 
+                    // Now copy the previously pre-scale row
                     for (var scaleY = 1; scaleY < scale; scaleY++)
                     {
                         Array.Copy(
                             target.Buffer,
                             targetRowStartIndex,
                             target.Buffer,
-                            targetRowStartIndex - rowByteLength,
+                            targetRowStartIndex + rowByteLength,
                             rowByteLength);
-                    
-                        targetRowStartIndex -= rowByteLength;
-                        targetByte1 -= rowByteLength;
+
+                        targetRowStartIndex += rowByteLength;
+                        targetByte1 += rowByteLength;
                     }
 
-                    targetRowStartIndex -= rowByteLength;
+                    targetRowStartIndex += targetWidth * GladeSelfRenderer.BytesPerPixel;
                 }
             }
         }
