+++ Glade2d/Graphics/SelfRenderer/BufferTransferring/Rotation270BufferTransferrer.cs
@@ -1,10 +1,9 @@
 ﻿using System;
-using Glade2d.Graphics.SelfRenderer;
 using Meadow.Foundation.Graphics.Buffers;
 
-namespace Glade2d.Graphics.BufferTransferring;
+namespace Glade2d.Graphics.SelfRenderer.BufferTransferring;
 
-public class Rotation90BufferTransferrer : IBufferTransferrer
+internal class Rotation270BufferTransferrer : IBufferTransferrer
 {
     public void Transfer(BufferRgb565 source, BufferRgb565 target, int scale)
     {
@@ -13,7 +12,6 @@ public class Rotation90BufferTransferrer : IBufferTransferrer
         var targetWidth = target.Width;
         var targetHeight = target.Height;
         var targetRowByteLength = targetWidth * GladeSelfRenderer.BytesPerPixel;
-        var sourceRowByteLength = sourceWidth * GladeSelfRenderer.BytesPerPixel;
 
         if (sourceWidth * scale != targetHeight || sourceHeight * scale != targetWidth)
         {
@@ -34,11 +32,10 @@ public class Rotation90BufferTransferrer : IBufferTransferrer
                 // rows. Writing to target in rows is important for Array.Copy
                 // calls for scaling.
 
-                // A 90 degree clockwise rotation means we read pixels column by 
-                // column and write them to the target row by row, starting from
-                // the last pixel in the row to the first.
-                var targetRowStartIndex = 0;
-                var targetByte1 = targetBufferPtr + targetRowByteLength - GladeSelfRenderer.BytesPerPixel;
+                // A 270 degree clockwise rotation means the source columns ascend while
+                // the target rows go backwards
+                var targetRowStartIndex = (targetHeight - 1) * targetWidth * GladeSelfRenderer.BytesPerPixel;
+                var targetByte1 = targetRowStartIndex + targetBufferPtr;
 
                 for (var sourceCol = 0; sourceCol < sourceWidth; sourceCol++)
                 {
@@ -50,10 +47,10 @@ public class Rotation90BufferTransferrer : IBufferTransferrer
                             *targetByte1 = *sourceByte1;
                             *(targetByte1 + 1) = *(sourceByte1 + 1);
 
-                            targetByte1 -= GladeSelfRenderer.BytesPerPixel;
+                            targetByte1 += GladeSelfRenderer.BytesPerPixel;
                         }
 
-                        sourceByte1 += sourceRowByteLength;
+                        sourceByte1 += sourceWidth * GladeSelfRenderer.BytesPerPixel;
                     }
 
                     for (var x = 1; x < scale; x++)
@@ -62,16 +59,15 @@ public class Rotation90BufferTransferrer : IBufferTransferrer
                             target.Buffer,
                             targetRowStartIndex,
                             target.Buffer,
-                            targetRowStartIndex + targetRowByteLength,
+                            targetRowStartIndex - targetRowByteLength,
                             targetRowByteLength);
 
-                        targetRowStartIndex += targetRowByteLength;
+                        targetRowStartIndex -= targetRowByteLength;
+                        targetByte1 -= targetRowByteLength;
                     }
 
-                    targetRowStartIndex += targetRowByteLength;
-
-                    // Go to the end of the row
-                    targetByte1 = targetBufferPtr + targetRowStartIndex + targetRowByteLength - GladeSelfRenderer.BytesPerPixel;
+                    targetRowStartIndex -= targetRowByteLength;
+                    targetByte1 = targetBufferPtr + targetRowStartIndex;
                 }
             }
         }
